Define worker inline to make it easier to bundle
To avoid references between threads and shared state, web workers have
to be started from a separate module. This creates a problem when the
application is to be bundled, if the name of the module is given as a
literal string: A module must exist with that name and must be
available at the same relative location to the calling code.

This can be problematic with `deno compile`: Not just must the
dispatcher be available in the work directory to pass to the
`--include` option, but in practice the entire library must be
vendored for the relative location to stay correct.

This patch attempt to address this problem. Since the dispatcher
includes no other modules and in less than 64kiB small, it can be
included as a text resource import and passed to the web worker as an
inline data: URL.

Using `--include` is then no longer necessary, and the library can be
used as any other. The flag `--unstable-raw-imports` must be used for
the time being, but this will probably graduate to stable in the
future.