Comment how to handle non-ASCII base64 encoding
If dispatcher.js contains a non-ASCII character, even just in a
comment, the btoa function will throw an error. It has been suggested
to use unescape(decodeURIComponent()) but this solution doesn't fix
the problem: For example it still raises an error for the expression:

    unescape(decodeURIComponent("let Блюм = 'їжа';"))

A better solution would have been to go through an Uint8Array, in the
form of:

    Buffer.from(dispatcher, 'UTF-8').toBase64()

This, however, works only in newer versions of Node and Deno, as the
toBase64 function was only recently added.

Since the dispatcher.js currently does not contain any problematic
characters, I have just added a comment about what could be done in
case this turns into a problem, not putting more requirement onto the
runtime than necessary.